+++ analyzer/src/main/java/nl/basjes/parse/useragent/analyze/treewalker/TreeExpressionEvaluator.java
@@ -67,6 +67,10 @@ public class TreeExpressionEvaluator implements Serializable {
         walkList = new WalkList(requiredPattern, matcher.getLookups(), matcher.getLookupSets(), verbose);
     }
 
+    public boolean isEmpty() {
+        return walkList.isEmpty();
+    }
+
     public void destroy() {
         walkList.destroy();
     }